Redesign Kext Download View
